Seperate calculateAverage function into its own utils file
