+++ package.json
@@ -42,6 +42,7 @@
   },
   "devDependencies": {
     "autoprefixer": "^10.4.7",
+    "cypress": "^9.6.1",
     "postcss": "^8.4.13",
     "tailwindcss": "^3.0.24"
   }
